requirements.txt
asgiref==3.8.1
Brotli==1.1.0
dj-database-url==2.1.0
Django==5.0.4
djangorestframework==3.14.0
gunicorn==22.0.0
packaging==24.0
pip==24.0
sqlparse==0.4.4
typing_extensions==4.11.0
whitenoise==6.6.0
django-cors-headers==4.4.0
pillow==10.4.0
requests == 2.32.3

